requirements.txt
# Core dependencies
praw==7.7.0
selenium==4.15.2
webdriver-manager==4.0.1
undetected-chromedriver==3.5.4
beautifulsoup4==4.12.2
requests==2.31.0

# Configuration management
pyyaml==6.0.1
python-dotenv==1.0.0

# NLP and text processing (simplified)
nltk==3.8.1
# Removing transformers and spacy due to compatibility issues
# transformers==4.30.2
# spacy==3.6.0

# Scheduling and time management
schedule==1.2.0
python-dateutil==2.8.2

# Utilities
tqdm==4.65.0
colorlog==6.7.0
retry==0.9.2

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Development tools
black==23.3.0
isort==5.12.0
flake8==6.0.0 